Added option for all tabs/organizations to use the same zoom factor.
Added functionality to synchronize zoom across all tabs and communities.

Moved the button above the spellchecker switch b ecause the  spellchecker   switch  has a subtitle associated with it that I don'tk now how to move so this makes more sense.

Set zoom to work across all tabs and not to affect  sidebar.